Fixes a bug where a blob couldn't be matched against twice.
The `ByteBuffer` `bb` here may be used multiple times.  The first time it is used, `bb` will correctly write out to the array `b`.  However, the second time it is used the buffer will already be fully read so `b` will be a zero length array with nothing in it.  By flipping the ByteBuffer after use, we are able to use it over and over again.